remove support for XEmacs
Some comments and doc-strings mentioning XEmacs are left as-is.
Likewise one regexp is left as-is, even though a comment mentions
that it is inefficient but has to be that way because of XEmacs.
It is unclear if that comment is correct and what one would want
to do when dropping XEmacs support.

Fixes #1.